Moving methods from Number object to GeneralNumber object
Part one.
